@@ -1,7 +1,7 @@
 
 package cz.habarta.typescript.generator;
 
-import cz.habarta.typescript.generator.util.Utils;
+import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.*;
 
@@ -16,11 +16,28 @@ public class CustomMappingTypeProcessor implements TypeProcessor {
 
     @Override
     public Result processType(Type javaType, Context context) {
-        final Class<?> rawClass = Utils.getRawClassOrNull(javaType);
-        if (rawClass != null) {
-            final String tsTypeName = customMappings.get(rawClass.getName());
+        if (javaType instanceof Class) {
+            final Class<?> javaClass = (Class<?>) javaType;
+            final String tsTypeName = customMappings.get(javaClass.getName());
             if (tsTypeName != null) {
                 return new Result(new TsType.BasicType(tsTypeName));
+            }
+        }
+        if (javaType instanceof ParameterizedType) {
+            final ParameterizedType parameterizedType = (ParameterizedType) javaType;
+            if (parameterizedType.getRawType() instanceof Class) {
+                final Class<?> javaClass = (Class<?>) parameterizedType.getRawType();
+                final String tsTypeName = customMappings.get(javaClass.getName());
+                if (tsTypeName != null) {
+                    final List<Class<?>> discoveredClasses = new ArrayList<>();
+                    final List<TsType> tsTypeArguments = new ArrayList<>();
+                    for (Type typeArgument : parameterizedType.getActualTypeArguments()) {
+                        final TypeProcessor.Result typeArgumentResult = context.processType(typeArgument);
+                        tsTypeArguments.add(typeArgumentResult.getTsType());
+                        discoveredClasses.addAll(typeArgumentResult.getDiscoveredClasses());
+                    }
+                    return new Result(new TsType.GenericBasicType(tsTypeName, tsTypeArguments), discoveredClasses);
+                }
             }
         }
         return null;
