@@ -1,15 +1,21 @@
 
 package cz.habarta.typescript.generator;
 
-import cz.habarta.typescript.generator.parser.*;
-import cz.habarta.typescript.generator.util.Predicate;
-import io.github.lukehutch.fastclasspathscanner.FastClasspathScanner;
-import io.github.lukehutch.fastclasspathscanner.scanner.ScanResult;
-import java.lang.reflect.*;
+import cz.habarta.typescript.generator.parser.SourceType;
+import cz.habarta.typescript.generator.util.Utils;
+import io.github.classgraph.ClassGraph;
+import io.github.classgraph.ScanResult;
+import java.lang.reflect.Type;
 import java.net.URLClassLoader;
-import java.util.*;
-import java.util.regex.Matcher;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
+import java.util.Objects;
+import java.util.function.Predicate;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 
 public class Input {
@@ -25,6 +31,7 @@ public class Input {
     }
 
     public static Input from(Type... types) {
+        Objects.requireNonNull(types, "types");
         final List<SourceType<Type>> sourceTypes = new ArrayList<>();
         for (Type type : types) {
             sourceTypes.add(new SourceType<>(type));
@@ -32,36 +39,81 @@ public class Input {
         return new Input(sourceTypes);
     }
 
-    public static Input fromClassNamesAndJaxrsApplication(List<String> classNames, List<String> classNamePatterns, String jaxrsApplicationClassName, boolean automaticJaxrsApplication, Predicate<String> isClassNameExcluded, URLClassLoader classLoader, boolean debug) {
+    public static Input fromClassNamesAndJaxrsApplication(List<String> classNames, List<String> classNamePatterns, String jaxrsApplicationClassName,
+            boolean automaticJaxrsApplication, Predicate<String> isClassNameExcluded, URLClassLoader classLoader, boolean debug) {
+        return fromClassNamesAndJaxrsApplication(classNames, classNamePatterns, null, null, null,
+            jaxrsApplicationClassName, automaticJaxrsApplication, isClassNameExcluded, classLoader,
+            debug);
+    }
+
+    public static Input fromClassNamesAndJaxrsApplication(List<String> classNames,
+        List<String> classNamePatterns, List<String> classesWithAnnotations,
+        List<String> classesImplementingInterfaces, List<String> classesExtendingClasses,
+        String jaxrsApplicationClassName,
+        boolean automaticJaxrsApplication, Predicate<String> isClassNameExcluded,
+        URLClassLoader classLoader, boolean debug) {
         final ClassLoader originalContextClassLoader = Thread.currentThread().getContextClassLoader();
         try {
-            Thread.currentThread().setContextClassLoader(classLoader);
-            final ClasspathScanner classpathScanner = new ClasspathScanner(classLoader, debug);
-            final List<SourceType<Type>> types = new ArrayList<>();
-            if (classNames != null) {
-                types.addAll(fromClassNames(classNames));
+            if (classLoader != null) {
+                Thread.currentThread().setContextClassLoader(classLoader);
             }
-            if (classNamePatterns != null) {
-                types.addAll(fromClassNamePatterns(classpathScanner.scanClasspath(), classNamePatterns));
+            try (final ClasspathScanner classpathScanner = new ClasspathScanner(classLoader, debug)) {
+                final List<SourceType<Type>> types = new ArrayList<>();
+                if (classNames != null) {
+                    types.addAll(fromClassNames(classNames));
+                }
+                if (classNamePatterns != null) {
+                    types.addAll(fromClassNamePatterns(classpathScanner.getScanResult(), classNamePatterns));
+                }
+                if (classesImplementingInterfaces != null) {
+                    final ScanResult scanResult = classpathScanner.getScanResult();
+                    final List<SourceType<Type>> c = fromClassNames(
+                        classesImplementingInterfaces.stream()
+                            .flatMap(interf -> scanResult.getClassesImplementing(interf).getNames()
+                                .stream())
+                            .distinct()
+                            .collect(Collectors.toList())
+                    );
+                    types.addAll(c);
+                }
+                if (classesExtendingClasses != null) {
+                    final ScanResult scanResult = classpathScanner.getScanResult();
+                    final List<SourceType<Type>> c = fromClassNames(
+                        classesExtendingClasses.stream()
+                            .flatMap(superclass -> scanResult.getSubclasses(superclass).getNames()
+                                .stream())
+                            .distinct()
+                            .collect(Collectors.toList())
+                    );
+                    types.addAll(c);
+                }
+                if (classesWithAnnotations != null) {
+                    final ScanResult scanResult = classpathScanner.getScanResult();
+                    types.addAll(fromClassNames(classesWithAnnotations.stream()
+                            .flatMap(annotation -> scanResult.getClassesWithAnnotation(annotation).getNames().stream())
+                            .distinct()
+                            .collect(Collectors.toList())
+                    ));
+                }
+                if (jaxrsApplicationClassName != null) {
+                    types.addAll(fromClassNames(Arrays.asList(jaxrsApplicationClassName)));
+                }
+                if (automaticJaxrsApplication) {
+                    types.addAll(JaxrsApplicationScanner.scanAutomaticJaxrsApplication(classpathScanner.getScanResult(), isClassNameExcluded));
+                }
+                if (types.isEmpty()) {
+                    final String errorMessage = "No input classes found.";
+                    TypeScriptGenerator.getLogger().error(errorMessage);
+                    throw new RuntimeException(errorMessage);
+                }
+                return new Input(types);
             }
-            if (jaxrsApplicationClassName != null) {
-                types.addAll(fromClassNames(Arrays.asList(jaxrsApplicationClassName)));
-            }
-            if (automaticJaxrsApplication) {
-                types.addAll(JaxrsApplicationScanner.scanAutomaticJaxrsApplication(classpathScanner.scanClasspath(), isClassNameExcluded));
-            }
-            if (types.isEmpty()) {
-                final String errorMessage = "No input classes found.";
-                System.out.println(errorMessage);
-                throw new RuntimeException(errorMessage);
-            }
-            return new Input(types);
         } finally {
             Thread.currentThread().setContextClassLoader(originalContextClassLoader);
         }
     }
 
-    private static class ClasspathScanner {
+    private static class ClasspathScanner implements AutoCloseable {
 
         private final URLClassLoader classLoader;
         private final boolean verbose;
@@ -72,32 +124,46 @@ public class Input {
             this.verbose = verbose;
         }
 
-        public ScanResult scanClasspath() {
+        public ScanResult getScanResult() {
             if (scanResult == null) {
-                System.out.println("Scanning classpath");
+                TypeScriptGenerator.getLogger().info("Scanning classpath");
                 final Date scanStart = new Date();
-                final ScanResult result = new FastClasspathScanner()
-                        .overrideClasspath((Object[])classLoader.getURLs())
-                        .verbose(verbose)
-                        .scan();
-                final int count = result.getNamesOfAllClasses().size();
+                ClassGraph classGraph = new ClassGraph()
+                        .enableClassInfo()
+                        .enableAnnotationInfo()
+                        .ignoreClassVisibility();
+                if (classLoader != null) {
+                    classGraph = classGraph.overrideClasspath((Object[])classLoader.getURLs());
+                }
+                if (verbose) {
+                    classGraph = classGraph.verbose();
+                }
+                final ScanResult result = classGraph.scan();
+                final int count = result.getAllClasses().size();
                 final Date scanEnd = new Date();
                 final double timeInSeconds = (scanEnd.getTime() - scanStart.getTime()) / 1000.0;
-                System.out.println(String.format("Scanning finished in %.2f seconds. Total number of classes: %d.", timeInSeconds, count));
+                TypeScriptGenerator.getLogger().info(String.format("Scanning finished in %.2f seconds. Total number of classes: %d.", timeInSeconds, count));
                 scanResult = result;
             }
             return scanResult;
+        }
+
+        @Override
+        public void close() {
+            if (scanResult != null) {
+                scanResult.close();
+            }
         }
 
     }
 
     private static List<SourceType<Type>> fromClassNamePatterns(ScanResult scanResult, List<String> classNamePatterns) {
         final List<String> allClassNames = new ArrayList<>();
-        allClassNames.addAll(scanResult.getNamesOfAllStandardClasses());
-        allClassNames.addAll(scanResult.getNamesOfAllInterfaceClasses());
+        allClassNames.addAll(scanResult.getAllStandardClasses().getNames());
+        allClassNames.addAll(scanResult.getAllInterfaces().getNames());
         Collections.sort(allClassNames);
         final List<String> classNames = filterClassNames(allClassNames, classNamePatterns);
-        System.out.println(String.format("Found %d classes matching pattern.", classNames.size()));
+        TypeScriptGenerator.getLogger().info(String.format("Found %d classes matching pattern.", classNames.size()));
         return fromClassNames(classNames);
     }
 
@@ -120,7 +186,7 @@ public class Input {
                 final Class<?> cls = Thread.currentThread().getContextClassLoader().loadClass(className);
                 classes.add(cls);
             } catch (ReflectiveOperationException e) {
-                System.out.println(String.format("Error: Cannot load class '%s'", className));
+                TypeScriptGenerator.getLogger().error(String.format("Cannot load class '%s'", className));
                 e.printStackTrace(System.out);
             }
         }
@@ -128,54 +194,14 @@ public class Input {
     }
 
     static List<String> filterClassNames(List<String> classNames, List<String> globs) {
-        final List<Pattern> regexps = globsToRegexps(globs);
+        final List<Pattern> regexps = Utils.globsToRegexps(globs);
         final List<String> result = new ArrayList<>();
         for (String className : classNames) {
-            if (classNameMatches(className, regexps)) {
+            if (Utils.classNameMatches(className, regexps)) {
                 result.add(className);
             }
         }
         return result;
-    }
-
-    static boolean classNameMatches(String className, List<Pattern> regexps) {
-        for (Pattern regexp : regexps) {
-            if (regexp.matcher(className).matches()) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    static List<Pattern> globsToRegexps(List<String> globs) {
-        final List<Pattern> regexps = new ArrayList<>();
-        for (String glob : globs) {
-            regexps.add(globToRegexp(glob));
-        }
-        return regexps;
-    }
-
-    /**
-     * Creates regexp for glob pattern.
-     * Replaces "*" with "[^.\$]*" and "**" with ".*".
-     */
-    static Pattern globToRegexp(String glob) {
-        final Pattern globToRegexpPattern = Pattern.compile("(\\*\\*)|(\\*)");
-        final Matcher matcher = globToRegexpPattern.matcher(glob);
-        final StringBuffer sb = new StringBuffer();
-        int lastEnd = 0;
-        while (matcher.find()) {
-            sb.append(Pattern.quote(glob.substring(lastEnd, matcher.start())));
-            if (matcher.group(1) != null) {
-                sb.append(Matcher.quoteReplacement(".*"));
-            }
-            if (matcher.group(2) != null) {
-                sb.append(Matcher.quoteReplacement("[^.$]*"));
-            }
-            lastEnd = matcher.end();
-        }
-        sb.append(Pattern.quote(glob.substring(lastEnd, glob.length())));
-        return Pattern.compile(sb.toString());
     }
 
 }
