@@ -7,18 +7,32 @@ import cz.habarta.typescript.generator.parser.*;
 import cz.habarta.typescript.generator.util.Utils;
 import java.io.*;
 import java.util.*;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 
 public class TypeScriptGenerator {
 
     public static final String Version = getVersion();
 
+    private static Logger logger = new Logger();
+
     private final Settings settings;
-    private TypeProcessor typeProcessor = null;
+    private TypeProcessor commonTypeProcessor = null;
     private ModelParser modelParser = null;
     private ModelCompiler modelCompiler = null;
     private Emitter emitter = null;
+    private InfoJsonEmitter infoJsonEmitter = null;
     private NpmPackageJsonEmitter npmPackageJsonEmitter = null;
+
+    public static Logger getLogger() {
+        return logger;
+    }
+
+    public static void setLogger(Logger logger) {
+        TypeScriptGenerator.logger = logger;
+    }
 
     public TypeScriptGenerator() {
         this (new Settings());
@@ -30,7 +44,7 @@ public class TypeScriptGenerator {
     }
 
     public static void printVersion() {
-        System.out.println("Running TypeScriptGenerator version " + Version);
+        TypeScriptGenerator.getLogger().info("Running TypeScriptGenerator version " + Version);
     }
 
     public String generateTypeScript(Input input) {
@@ -43,6 +57,7 @@ public class TypeScriptGenerator {
         generateTypeScript(input, output, false, 0);
     }
 
+    @Deprecated
     public void generateEmbeddableTypeScript(Input input, Output output, boolean addExportKeyword, int initialIndentationLevel) {
         generateTypeScript(input, output, addExportKeyword, initialIndentationLevel);
     }
@@ -50,8 +65,24 @@ public class TypeScriptGenerator {
     private void generateTypeScript(Input input, Output output, boolean forceExportKeyword, int initialIndentationLevel) {
         final Model model = getModelParser().parseModel(input.getSourceTypes());
         final TsModel tsModel = getModelCompiler().javaToTypeScript(model);
-        getEmitter().emit(tsModel, output.getWriter(), output.getName(), output.shouldCloseWriter(), forceExportKeyword, initialIndentationLevel);
+        generateTypeScript(tsModel, output, forceExportKeyword, initialIndentationLevel);
+        generateInfoJson(tsModel, output);
         generateNpmPackageJson(output);
+    }
+
+    private void generateTypeScript(TsModel tsModel, Output output, boolean forceExportKeyword, int initialIndentationLevel) {
+        getEmitter().emit(tsModel, output.getWriter(), output.getName(), output.shouldCloseWriter(), forceExportKeyword, initialIndentationLevel);
+    }
+
+    private void generateInfoJson(TsModel tsModel, Output output) {
+        if (settings.generateInfoJson) {
+            if (output.getName() == null) {
+                throw new RuntimeException("Generating info JSON can only be used when output is specified using file name");
+            }
+            final File outputFile = new File(output.getName());
+            final Output out = Output.to(new File(outputFile.getParent(), "typescript-generator-info.json"));
+            getInfoJsonEmitter().emit(tsModel, out.getWriter(), out.getName(), out.shouldCloseWriter());
+        }
     }
 
     private void generateNpmPackageJson(Output output) {
@@ -65,27 +96,56 @@ public class TypeScriptGenerator {
             npmPackageJson.name = settings.npmName;
             npmPackageJson.version = settings.npmVersion;
             npmPackageJson.types = outputFile.getName();
+            npmPackageJson.dependencies = new LinkedHashMap<>();
+            if (settings.moduleDependencies != null) {
+                for (ModuleDependency dependency : settings.moduleDependencies) {
+                    npmPackageJson.dependencies.put(dependency.npmPackageName, dependency.npmVersionRange);
+                }
+            }
             if (settings.outputFileType == TypeScriptFileType.implementationFile) {
+                npmPackageJson.types = Utils.replaceExtension(outputFile, ".d.ts").getName();
                 npmPackageJson.main = Utils.replaceExtension(outputFile, ".js").getName();
-                npmPackageJson.dependencies = !settings.npmPackageDependencies.isEmpty() ? settings.npmPackageDependencies : null;
+                npmPackageJson.dependencies.putAll(settings.npmPackageDependencies);
                 npmPackageJson.devDependencies = Collections.singletonMap("typescript", settings.typescriptVersion);
-                npmPackageJson.scripts = Collections.singletonMap("build", "tsc --module umd --moduleResolution node --sourceMap " + outputFile.getName());
+                final String npmBuildScript = settings.npmBuildScript != null
+                        ? settings.npmBuildScript
+                        : "tsc --module umd --moduleResolution node --target es5 --lib es6 --declaration --sourceMap $outputFile";
+                final String build = npmBuildScript.replaceAll(Pattern.quote("$outputFile"), outputFile.getName());
+                npmPackageJson.scripts = Collections.singletonMap("build", build);
+            }
+            if (npmPackageJson.dependencies.isEmpty()) {
+                npmPackageJson.dependencies = null;
             }
             getNpmPackageJsonEmitter().emit(npmPackageJson, npmOutput.getWriter(), npmOutput.getName(), npmOutput.shouldCloseWriter());
         }
     }
 
-    public TypeProcessor getTypeProcessor() {
-        if (typeProcessor == null) {
-            final List<TypeProcessor> processors = new ArrayList<>();
-            processors.add(new ExcludingTypeProcessor(settings.getExcludeFilter()));
-            if (settings.customTypeProcessor != null) {
-                processors.add(settings.customTypeProcessor);
-            }
-            processors.add(new CustomMappingTypeProcessor(settings.customTypeMappings));
-            processors.add(new DefaultTypeProcessor());
-            typeProcessor = new TypeProcessor.Chain(processors);
+    public TypeProcessor getCommonTypeProcessor() {
+        if (commonTypeProcessor == null) {
+            final List<RestApplicationParser.Factory> restFactories = settings.getRestApplicationParserFactories();
+            final ModelParser.Factory modelParserFactory = getModelParserFactory();
+            final List<TypeProcessor> specificTypeProcessors = Stream
+                    .concat(
+                            restFactories.stream().map(factory -> factory.getSpecificTypeProcessor()),
+                            Stream.of(modelParserFactory.getSpecificTypeProcessor())
+                    )
+                    .filter(Objects::nonNull)
+                    .collect(Collectors.toList());
+            commonTypeProcessor = createTypeProcessor(specificTypeProcessors);
         }
+        return commonTypeProcessor;
+    }
+
+    private TypeProcessor createTypeProcessor(List<TypeProcessor> specificTypeProcessors) {
+        final List<TypeProcessor> processors = new ArrayList<>();
+        processors.add(new ExcludingTypeProcessor(settings.getExcludeFilter()));
+        if (settings.customTypeProcessor != null) {
+            processors.add(settings.customTypeProcessor);
+        }
+        processors.add(new CustomMappingTypeProcessor(settings.customTypeMappings));
+        processors.addAll(specificTypeProcessors);
+        processors.add(new DefaultTypeProcessor());
+        final TypeProcessor typeProcessor = new TypeProcessor.Chain(processors);
         return typeProcessor;
     }
 
@@ -97,13 +157,21 @@ public class TypeScriptGenerator {
     }
 
     private ModelParser createModelParser() {
+        final List<RestApplicationParser.Factory> factories = settings.getRestApplicationParserFactories();
+        final List<RestApplicationParser> restApplicationParsers = factories.stream()
+                .map(factory -> factory.create(settings, getCommonTypeProcessor()))
+                .collect(Collectors.toList());
+        return getModelParserFactory().create(settings, getCommonTypeProcessor(), restApplicationParsers);
+    }
+
+    private ModelParser.Factory getModelParserFactory() {
         switch (settings.jsonLibrary) {
             case jackson1:
-                return new Jackson1Parser(settings, getTypeProcessor());
+                return new Jackson1Parser.Factory();
             case jackson2:
-                return new Jackson2Parser(settings, getTypeProcessor());
+                return new Jackson2Parser.Jackson2ParserFactory();
             case jaxb:
-                return new Jackson2Parser(settings, getTypeProcessor(), /*useJaxbAnnotations*/ true);
+                return new Jackson2Parser.JaxbParserFactory();
             default:
                 throw new RuntimeException();
         }
@@ -111,7 +179,7 @@ public class TypeScriptGenerator {
 
     public ModelCompiler getModelCompiler() {
         if (modelCompiler == null) {
-            modelCompiler = new ModelCompiler(settings, getTypeProcessor());
+            modelCompiler = new ModelCompiler(settings, getCommonTypeProcessor());
         }
         return modelCompiler;
     }
@@ -121,6 +189,13 @@ public class TypeScriptGenerator {
             emitter = new Emitter(settings);
         }
         return emitter;
+    }
+
+    public InfoJsonEmitter getInfoJsonEmitter() {
+        if (infoJsonEmitter == null) {
+            infoJsonEmitter = new InfoJsonEmitter();
+        }
+        return infoJsonEmitter;
     }
 
     public NpmPackageJsonEmitter getNpmPackageJsonEmitter() {
