@@ -12,19 +12,59 @@ public class TsBeanModel extends TsDeclarationModel {
     private final boolean isClass;
     private final List<TsType.GenericVariableType> typeParameters;
     private final TsType parent;
+    private final List<TsType> extendsList;
+    private final List<TsType> implementsList;
     private final List<Class<?>> taggedUnionClasses;
-    private final List<TsType> interfaces;
+    private final String discriminantProperty;
+    private final String discriminantLiteral;
+    private final TsAliasModel taggedUnionAlias;
     private final List<TsPropertyModel> properties;
     private final TsConstructorModel constructor;
     private final List<TsMethodModel> methods;
 
-    public TsBeanModel(Class<?> origin, TsBeanCategory category, boolean isClass, Symbol name, List<TsType.GenericVariableType> typeParameters, TsType parent, List<Class<?>> taggedUnionClasses, List<TsType> interfaces, List<TsPropertyModel> properties, TsConstructorModel constructor, List<TsMethodModel> methods, List<String> comments) {
+    public TsBeanModel(
+            Class<?> origin,
+            TsBeanCategory category,
+            boolean isClass,
+            Symbol name,
+            List<TsType.GenericVariableType> typeParameters,
+            TsType parent,
+            List<TsType> extendsList,
+            List<TsType> implementsList,
+            List<TsPropertyModel> properties,
+            TsConstructorModel constructor,
+            List<TsMethodModel> methods,
+            List<String> comments) {
+        this(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, null, null, null, null, properties, constructor, methods, comments);
+    }
+
+    private TsBeanModel(
+            Class<?> origin,
+            TsBeanCategory category,
+            boolean isClass,
+            Symbol name,
+            List<TsType.GenericVariableType> typeParameters,
+            TsType parent,
+            List<TsType> extendsList,
+            List<TsType> implementsList,
+            List<Class<?>> taggedUnionClasses,
+            String discriminantProperty,
+            String discriminantLiteral,
+            TsAliasModel taggedUnionAlias,
+            List<TsPropertyModel> properties,
+            TsConstructorModel constructor,
+            List<TsMethodModel> methods,
+            List<String> comments) {
         super(origin, category, name, comments);
         this.isClass = isClass;
         this.typeParameters = Utils.listFromNullable(typeParameters);
         this.parent = parent;
+        this.extendsList = Utils.listFromNullable(extendsList);
+        this.implementsList = Utils.listFromNullable(implementsList);
         this.taggedUnionClasses = Utils.listFromNullable(taggedUnionClasses);
-        this.interfaces = Utils.listFromNullable(interfaces);
+        this.discriminantProperty = discriminantProperty;
+        this.discriminantLiteral = discriminantLiteral;
+        this.taggedUnionAlias = taggedUnionAlias;
         this.properties = Utils.listFromNullable(properties);
         this.constructor = constructor;
         this.methods = Utils.listFromNullable(methods);
@@ -42,33 +82,43 @@ public class TsBeanModel extends TsDeclarationModel {
         return parent;
     }
 
+    public List<TsType> getExtendsList() {
+        return extendsList;
+    }
+
+    public List<TsType> getImplementsList() {
+        return implementsList;
+    }
+
+    public List<TsType> getAllParents() {
+        final List<TsType> parents = new ArrayList<>();
+        parents.addAll(extendsList);
+        parents.addAll(implementsList);
+        return parents;
+    }
+
     public List<Class<?>> getTaggedUnionClasses() {
         return taggedUnionClasses;
     }
 
-    public List<TsType> getInterfaces() {
-        return interfaces;
+    public String getDiscriminantProperty() {
+        return discriminantProperty;
     }
 
-    public List<TsType> getParentAndInterfaces() {
-        final List<TsType> parents = new ArrayList<>();
-        if (parent != null) {
-            parents.add(parent);
-        }
-        parents.addAll(interfaces);
-        return parents;
+    public String getDiscriminantLiteral() {
+        return discriminantLiteral;
     }
 
-    public List<TsType> getExtendsList() {
-        return isClass
-                ? Utils.listFromNullable(parent)
-                : getParentAndInterfaces();
+    public TsBeanModel withTaggedUnion(List<Class<?>> taggedUnionClasses, String discriminantProperty, String discriminantLiteral) {
+        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, taggedUnionClasses, discriminantProperty, discriminantLiteral, taggedUnionAlias, properties, constructor, methods, comments);
     }
 
-    public List<TsType> getImplementsList() {
-        return isClass
-                ? interfaces
-                : Collections.<TsType>emptyList();
+    public TsAliasModel getTaggedUnionAlias() {
+        return taggedUnionAlias;
+    }
+
+    public TsBeanModel withTaggedUnionAlias(TsAliasModel taggedUnionAlias) {
+        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, taggedUnionClasses, discriminantProperty, discriminantLiteral, taggedUnionAlias, properties, constructor, methods, comments);
     }
 
     public List<TsPropertyModel> getProperties() {
@@ -76,11 +126,15 @@ public class TsBeanModel extends TsDeclarationModel {
     }
 
     public TsBeanModel withProperties(List<TsPropertyModel> properties) {
-        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, taggedUnionClasses, interfaces, properties, constructor, methods, comments);
+        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, taggedUnionClasses, discriminantProperty, discriminantLiteral, taggedUnionAlias, properties, constructor, methods, comments);
     }
 
     public TsConstructorModel getConstructor() {
         return constructor;
+    }
+
+    public TsBeanModel withConstructor(TsConstructorModel constructor) {
+        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, taggedUnionClasses, discriminantProperty, discriminantLiteral, taggedUnionAlias, properties, constructor, methods, comments);
     }
 
     public List<TsMethodModel> getMethods() {
@@ -88,11 +142,15 @@ public class TsBeanModel extends TsDeclarationModel {
     }
 
     public TsBeanModel withMethods(List<TsMethodModel> methods) {
-        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, taggedUnionClasses, interfaces, properties, constructor, methods, comments);
+        return new TsBeanModel(origin, category, isClass, name, typeParameters, parent, extendsList, implementsList, taggedUnionClasses, discriminantProperty, discriminantLiteral, taggedUnionAlias, properties, constructor, methods, comments);
     }
 
     public boolean isJaxrsApplicationClientBean() {
-        return category== TsBeanCategory.Service && isClass;
+        return category == TsBeanCategory.Service && isClass;
+    }
+    
+    public boolean isDataClass() {
+        return category == TsBeanCategory.Data && isClass;
     }
     
 }
