@@ -6,8 +6,8 @@ import java.util.List;
 
 public class TsConstructorModel extends TsCallableModel {
     
-    public TsConstructorModel(List<TsParameterModel> parameters, List<TsStatement> body, List<String> comments) {
-        super("constructor", null, parameters, body, comments);
+    public TsConstructorModel(TsModifierFlags modifiers, List<TsParameterModel> parameters, List<TsStatement> body, List<String> comments) {
+        super("constructor", modifiers, null, parameters, null, body, comments);
     }
 
 }
