@@ -12,15 +12,22 @@ public class TsEnumModel extends TsDeclarationModel {
 
     private final EnumKind kind;
     private final List<EnumMemberModel> members;
+    private final boolean isNonConstEnum;
 
+    @Deprecated
     public TsEnumModel(Class<?> origin, Symbol name, EnumKind kind, List<EnumMemberModel> members, List<String> comments) {
+        this(origin, name, kind, members, comments, false);
+    }
+
+    public TsEnumModel(Class<?> origin, Symbol name, EnumKind kind, List<EnumMemberModel> members, List<String> comments, boolean isNonConstEnum) {
         super(origin, null, name, comments);
         this.kind = kind;
         this.members = members;
+        this.isNonConstEnum = isNonConstEnum;
     }
 
-    public static TsEnumModel fromEnumModel(Symbol name, EnumModel enumModel) {
-        return new TsEnumModel(enumModel.getOrigin(), name, enumModel.getKind(), enumModel.getMembers(), enumModel.getComments());
+    public static TsEnumModel fromEnumModel(Symbol name, EnumModel enumModel, boolean isNonConstEnum) {
+        return new TsEnumModel(enumModel.getOrigin(), name, enumModel.getKind(), enumModel.getMembers(), enumModel.getComments(), isNonConstEnum);
     }
 
     public EnumKind getKind() {
@@ -31,8 +38,12 @@ public class TsEnumModel extends TsDeclarationModel {
         return members;
     }
 
+    public boolean isNonConstEnum() {
+        return isNonConstEnum;
+    }
+
     public TsEnumModel withMembers(List<EnumMemberModel> members) {
-        return new TsEnumModel(origin, name, kind, members, comments);
+        return new TsEnumModel(origin, name, kind, members, comments, isNonConstEnum);
     }
 
 }
