@@ -7,8 +7,8 @@ import java.util.List;
 
 public class TsMethodModel extends TsCallableModel {
     
-    public TsMethodModel(String name, TsType returnType, List<TsParameterModel> parameters, List<TsStatement> body, List<String> comments) {
-        super(name, returnType, parameters, body, comments);
+    public TsMethodModel(String name, TsModifierFlags modifiers, List<TsType.GenericVariableType> typeParameters, List<TsParameterModel> parameters, TsType returnType, List<TsStatement> body, List<String> comments) {
+        super(name, modifiers, typeParameters, parameters, returnType, body, comments);
     }
 
 }
