@@ -1,8 +1,8 @@
 
 package cz.habarta.typescript.generator.parser;
 
-import cz.habarta.typescript.generator.compiler.EnumMemberModel;
 import cz.habarta.typescript.generator.compiler.EnumKind;
+import cz.habarta.typescript.generator.compiler.EnumMemberModel;
 import java.util.*;
 
 
