@@ -3,16 +3,29 @@ package cz.habarta.typescript.generator.parser;
 
 import cz.habarta.typescript.generator.JaxrsApplicationScanner;
 import cz.habarta.typescript.generator.Settings;
-import cz.habarta.typescript.generator.util.Parameter;
-import cz.habarta.typescript.generator.util.Predicate;
+import cz.habarta.typescript.generator.TsType;
+import cz.habarta.typescript.generator.TypeProcessor;
+import cz.habarta.typescript.generator.TypeScriptGenerator;
 import cz.habarta.typescript.generator.util.Utils;
+import java.beans.BeanInfo;
+import java.beans.IntrospectionException;
+import java.beans.Introspector;
+import java.beans.PropertyDescriptor;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.lang.reflect.Parameter;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Stream;
 import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.BeanParam;
 import javax.ws.rs.CookieParam;
 import javax.ws.rs.FormParam;
 import javax.ws.rs.HeaderParam;
@@ -21,6 +34,7 @@ import javax.ws.rs.MatrixParam;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.container.Suspended;
 import javax.ws.rs.core.Application;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.GenericEntity;
@@ -28,35 +42,42 @@ import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.StreamingOutput;
 
+public class JaxrsApplicationParser extends RestApplicationParser {
 
-public class JaxrsApplicationParser {
+    public static class Factory extends RestApplicationParser.Factory {
 
-    private final Settings settings;
-    private final Predicate<String> isClassNameExcluded;
-    private final Set<String> defaultExcludes;
-    private final JaxrsApplicationModel model;
-
-    public JaxrsApplicationParser(Settings settings) {
-        this.settings = settings;
-        this.isClassNameExcluded = settings.getExcludeFilter();
-        this.defaultExcludes = new LinkedHashSet<>(getDefaultExcludedClassNames());
-        this.model = new JaxrsApplicationModel();
-    }
-
-    public JaxrsApplicationModel getModel() {
-        return model;
-    }
-
-    public static class Result {
-        public List<SourceType<Type>> discoveredTypes;
-        public Result() {
-            discoveredTypes = new ArrayList<>();
+        @Override
+        public TypeProcessor getSpecificTypeProcessor() {
+            return (javaType, context) -> {
+                final Class<?> rawClass = Utils.getRawClassOrNull(javaType);
+                if (rawClass != null) {
+                    for (Map.Entry<Class<?>, TsType> entry : getStandardEntityClassesMapping().entrySet()) {
+                        final Class<?> cls = entry.getKey();
+                        final TsType type = entry.getValue();
+                        if (cls.isAssignableFrom(rawClass)) {
+                            return type != null ? new TypeProcessor.Result(type) : null;
+                        }
+                    }
+                    if (getDefaultExcludedClassNames().contains(rawClass.getName())) {
+                        return new TypeProcessor.Result(TsType.Any);
+                    }
+                }
+                return null;
+            };
         }
-        public Result(List<SourceType<Type>> discoveredTypes) {
-            this.discoveredTypes = discoveredTypes;
+
+        @Override
+        public JaxrsApplicationParser create(Settings settings, TypeProcessor commonTypeProcessor) {
+            return new JaxrsApplicationParser(settings, commonTypeProcessor);
         }
+
+    };
+
+    public JaxrsApplicationParser(Settings settings, TypeProcessor commonTypeProcessor) {
+        super(settings, commonTypeProcessor, new RestApplicationModel(RestApplicationType.Jaxrs));
     }
 
+    @Override
     public Result tryParse(SourceType<?> sourceType) {
         if (!(sourceType.type instanceof Class<?>)) {
             return null;
@@ -77,7 +98,7 @@ public class JaxrsApplicationParser {
         // resource
         final Path path = cls.getAnnotation(Path.class);
         if (path != null) {
-            System.out.println("Parsing JAX-RS resource: " + cls.getName());
+            TypeScriptGenerator.getLogger().verbose("Parsing JAX-RS resource: " + cls.getName());
             final Result result = new Result();
             parseResource(result, new ResourceContext(cls, path.value()), cls);
             return result;
@@ -98,20 +119,7 @@ public class JaxrsApplicationParser {
         final ResourceContext subContext = context.subPathParamTypes(pathParamTypes);
         // parse resource methods
         final List<Method> methods = Arrays.asList(resourceClass.getMethods());
-        Collections.sort(methods, new Comparator<Method>() {
-            @Override
-            public int compare(Method o1, Method o2) {
-                final int nameDiff = o1.getName().compareToIgnoreCase(o2.getName());
-                if (nameDiff != 0) {
-                    return nameDiff;
-                }
-                final int parameterTypesDiff = Arrays.asList(o1.getParameterTypes()).toString().compareTo(Arrays.asList(o2.getParameterTypes()).toString());
-                if (parameterTypesDiff != 0) {
-                    return parameterTypesDiff;
-                }
-                return 0;
-            }
-        });
+        Collections.sort(methods, Utils.methodComparator());
         for (Method method : methods) {
             parseResourceMethod(result, subContext, resourceClass, method);
         }
@@ -120,10 +128,9 @@ public class JaxrsApplicationParser {
     private void parseResourceMethod(Result result, ResourceContext context, Class<?> resourceClass, Method method) {
         final Path pathAnnotation = method.getAnnotation(Path.class);
         // subContext
-        context = context.subPath(pathAnnotation);
+        context = context.subPath(pathAnnotation != null ? pathAnnotation.value() : null);
         final Map<String, Type> pathParamTypes = new LinkedHashMap<>();
-        final List<Parameter> methodParameters = Parameter.ofMethod(method);
-        for (Parameter parameter : methodParameters) {
+        for (Parameter parameter : method.getParameters()) {
             final PathParam pathParamAnnotation = parameter.getAnnotation(PathParam.class);
             if (pathParamAnnotation != null) {
                 pathParamTypes.put(pathParamAnnotation.value(), parameter.getParameterizedType());
@@ -153,17 +160,30 @@ public class JaxrsApplicationParser {
             for (PathTemplate.Part part : pathTemplate.getParts()) {
                 if (part instanceof PathTemplate.Parameter) {
                     final PathTemplate.Parameter parameter = (PathTemplate.Parameter) part;
-                    final Type type = context.pathParamTypes.get(parameter.getName());
-                    pathParams.add(new MethodParameterModel(parameter.getName(), type != null ? type : String.class));
+                    final Type type = context.pathParamTypes.get(parameter.getOriginalName());
+                    final Type paramType = type != null ? type : String.class;
+                    pathParams.add(new MethodParameterModel(parameter.getValidName(), paramType));
+                    foundType(result, paramType, resourceClass, method.getName());
                 }
             }
             // query parameters
-            final List<MethodParameterModel> queryParams = new ArrayList<>();
-            final List<Parameter> params = Parameter.ofMethod(method);
-            for (Parameter param : params) {
+            final List<RestQueryParam> queryParams = new ArrayList<>();
+            for (Parameter param : method.getParameters()) {
                 final QueryParam queryParamAnnotation = param.getAnnotation(QueryParam.class);
                 if (queryParamAnnotation != null) {
-                    queryParams.add(new MethodParameterModel(queryParamAnnotation.value(), param.getParameterizedType()));
+                    queryParams.add(new RestQueryParam.Single(new MethodParameterModel(queryParamAnnotation.value(), param.getParameterizedType())));
+                    foundType(result, param.getParameterizedType(), resourceClass, method.getName());
+                }
+                final BeanParam beanParamAnnotation = param.getAnnotation(BeanParam.class);
+                if (beanParamAnnotation != null) {
+                    final Class<?> beanParamClass = param.getType();
+                    final BeanModel paramBean = getQueryParameters(beanParamClass);
+                    if (paramBean != null) {
+                        queryParams.add(new RestQueryParam.Bean(paramBean));
+                        for (PropertyModel property : paramBean.getProperties()) {
+                            foundType(result, property.getType(), beanParamClass, property.getName());
+                        }
+                    }
                 }
             }
             // JAX-RS specification - 3.3.2.1 Entity Parameters
@@ -176,7 +196,17 @@ public class JaxrsApplicationParser {
             final Type genericReturnType = method.getGenericReturnType();
             final Type modelReturnType;
             if (returnType == void.class) {
-                modelReturnType = returnType;
+                //for async response also use swagger
+                if (hasAnyAnnotation(method.getParameters(), Collections.singletonList(Suspended.class))) {
+                    if (swaggerOperation.responseType != null) {
+                        modelReturnType = swaggerOperation.responseType;
+                        foundType(result, modelReturnType, resourceClass, method.getName());
+                    } else {
+                        modelReturnType = Object.class;
+                    }
+                } else {
+                    modelReturnType = returnType;
+                }
             } else if (returnType == Response.class) {
                 if (swaggerOperation.responseType != null) {
                     modelReturnType = swaggerOperation.responseType;
@@ -195,39 +225,13 @@ public class JaxrsApplicationParser {
             // comments
             final List<String> comments = Swagger.getOperationComments(swaggerOperation);
             // create method
-            model.getMethods().add(new JaxrsMethodModel(resourceClass, method.getName(), modelReturnType,
+            model.getMethods().add(new RestMethodModel(resourceClass, method.getName(), modelReturnType,
                     context.rootResource, httpMethod.value(), context.path, pathParams, queryParams, entityParameter, comments));
         }
         // JAX-RS specification - 3.4.1 Sub Resources
         if (pathAnnotation != null && httpMethod == null) {
             parseResource(result, context, method.getReturnType());
         }
-    }
-
-    private void foundType(Result result, Type type, Class<?> usedInClass, String usedInMember) {
-        if (!isExcluded(type)) {
-            result.discoveredTypes.add(new SourceType<>(type, usedInClass, usedInMember));
-        }
-    }
-
-    private boolean isExcluded(Type type) {
-        final Class<?> cls = Utils.getRawClassOrNull(type);
-        if (cls == null) {
-            return false;
-        }
-        if (isClassNameExcluded != null && isClassNameExcluded.test(cls.getName())) {
-            return true;
-        }
-        if (defaultExcludes.contains(cls.getName())) {
-            return true;
-        }
-
-        for (Class<?> standardEntityClass : getStandardEntityClasses()) {
-            if (standardEntityClass.isAssignableFrom(cls)) {
-                return true;
-            }
-        }
-        return false;
     }
 
     private static HttpMethod getHttpMethod(Method method) {
@@ -240,86 +244,97 @@ public class JaxrsApplicationParser {
         return null;
     }
 
+    private static BeanModel getQueryParameters(Class<?> paramBean) {
+        final List<PropertyModel> properties = new ArrayList<>();
+        final List<Field> fields = Utils.getAllFields(paramBean);
+        for (Field field : fields) {
+            final QueryParam annotation = field.getAnnotation(QueryParam.class);
+            if (annotation != null) {
+                properties.add(new PropertyModel(annotation.value(), field.getGenericType(), /*optional*/true, field, null, null, null));
+            }
+        }
+        try {
+            final BeanInfo beanInfo = Introspector.getBeanInfo(paramBean);
+            for (PropertyDescriptor propertyDescriptor : beanInfo.getPropertyDescriptors()) {
+                final Method writeMethod = propertyDescriptor.getWriteMethod();
+                if (writeMethod != null) {
+                    final QueryParam annotation = writeMethod.getAnnotation(QueryParam.class);
+                    if (annotation != null) {
+                        properties.add(new PropertyModel(annotation.value(), propertyDescriptor.getPropertyType(), /*optional*/true, writeMethod, null, null, null));
+                    }
+                }
+            }
+        } catch (IntrospectionException e) {
+            TypeScriptGenerator.getLogger().warning(String.format("Cannot introspect '%s' class: " + e.getMessage(), paramBean));
+        }
+        if (properties.isEmpty()) {
+            return null;
+        } else {
+            return new BeanModel(paramBean, null, null, null, null, null, properties, null);
+        }
+    }
+
     private static MethodParameterModel getEntityParameter(Method method) {
-        final List<Parameter> parameters = Parameter.ofMethod(method);
-        for (Parameter parameter : parameters) {
-            if (!hasAnyAnnotation(parameter, Arrays.asList(
+        for (Parameter parameter : method.getParameters()) {
+            if (!Utils.hasAnyAnnotation(parameter::getAnnotation, Arrays.asList(
                     MatrixParam.class,
                     QueryParam.class,
                     PathParam.class,
                     CookieParam.class,
                     HeaderParam.class,
+                    Suspended.class,
                     Context.class,
-                    FormParam.class
-                    ))) {
+                    FormParam.class,
+                    BeanParam.class
+            ))) {
                 return new MethodParameterModel(parameter.getName(), parameter.getParameterizedType());
             }
         }
         return null;
     }
-
-    private static boolean hasAnyAnnotation(Parameter parameter, List<Class<? extends Annotation>> annotationClasses) {
-        for (Class<? extends Annotation> annotationClass : annotationClasses) {
-            for (Annotation parameterAnnotation : parameter.getAnnotations()) {
-                if (annotationClass.isInstance(parameterAnnotation)) {
-                    return true;
-                }
-            }
-        }
-        return false;
+    
+    private static boolean hasAnyAnnotation(Parameter[] parameters, List<Class<? extends Annotation>> annotationClasses) {
+        return Stream.of(parameters)
+                .anyMatch(parameter -> Utils.hasAnyAnnotation(parameter::getAnnotation, annotationClasses));
     }
-
-    public static List<Class<?>> getStandardEntityClasses() {
+    
+    private static Map<Class<?>, TsType> getStandardEntityClassesMapping() {
         // JAX-RS specification - 4.2.4 Standard Entity Providers
-        return Arrays.asList(
-                byte[].class,
-                java.lang.String.class,
-                java.io.InputStream.class,
-                java.io.Reader.class,
-                java.io.File.class,
-                javax.activation.DataSource.class,
-                javax.xml.transform.Source.class,
-                javax.xml.bind.JAXBElement.class,
-                MultivaluedMap.class,
-                StreamingOutput.class,
-                java.lang.Boolean.class, java.lang.Character.class, java.lang.Number.class,
-                long.class, int.class, short.class, byte.class, double.class, float.class, boolean.class, char.class);
+        if (standardEntityClassesMapping == null) {
+            final Map<Class<?>, TsType> map = new LinkedHashMap<>();
+            // null value means that class is handled by DefaultTypeProcessor
+            map.put(byte[].class, TsType.Any);
+            map.put(java.lang.String.class, null);
+            map.put(java.io.InputStream.class, TsType.Any);
+            map.put(java.io.Reader.class, TsType.Any);
+            map.put(java.io.File.class, TsType.Any);
+            map.put(javax.activation.DataSource.class, TsType.Any);
+            map.put(javax.xml.transform.Source.class, TsType.Any);
+            map.put(javax.xml.bind.JAXBElement.class, null);
+            map.put(MultivaluedMap.class, TsType.Any);
+            map.put(StreamingOutput.class, TsType.Any);
+            map.put(java.lang.Boolean.class, null);
+            map.put(java.lang.Character.class, null);
+            map.put(java.lang.Number.class, null);
+            map.put(long.class, null);
+            map.put(int.class, null);
+            map.put(short.class, null);
+            map.put(byte.class, null);
+            map.put(double.class, null);
+            map.put(float.class, null);
+            map.put(boolean.class, null);
+            map.put(char.class, null);
+            standardEntityClassesMapping = map;
+        }
+        return standardEntityClassesMapping;
     }
+
+    private static Map<Class<?>, TsType> standardEntityClassesMapping;
 
     private static List<String> getDefaultExcludedClassNames() {
         return Arrays.asList(
                 "org.glassfish.jersey.media.multipart.FormDataBodyPart"
         );
-    }
-
-    private static class ResourceContext {
-        public final Class<?> rootResource;
-        public final String path;
-        public final Map<String, Type> pathParamTypes;
-
-        public ResourceContext(Class<?> rootResource, String path) {
-            this(rootResource, path, new LinkedHashMap<String, Type>());
-        }
-
-        private ResourceContext(Class<?> rootResource, String path, Map<String, Type> pathParamTypes) {
-            this.rootResource = rootResource;
-            this.path = path;
-            this.pathParamTypes = pathParamTypes;
-        }
-
-        ResourceContext subPath(Path pathAnnotation) {
-            final String subPath = pathAnnotation != null ? pathAnnotation.value() : null;
-            return new ResourceContext(rootResource, Utils.joinPath(path, subPath), pathParamTypes);
-        }
-
-        ResourceContext subPathParamTypes(Map<String, Type> subPathParamTypes) {
-            final Map<String, Type> newPathParamTypes = new LinkedHashMap<>();
-            newPathParamTypes.putAll(pathParamTypes);
-            if (subPathParamTypes != null) {
-                newPathParamTypes.putAll(subPathParamTypes);
-            }
-            return new ResourceContext(rootResource, path, newPathParamTypes);
-        }
     }
 
 }
