@@ -13,7 +13,7 @@ class AxiosHttpClient implements HttpClient<Axios.AxiosRequestConfig> {
     constructor(private axios: Axios.AxiosInstance) {
     }
 
-    request(requestConfig: { method: string; url: string; queryParams?: any; data?: any; options?: Axios.AxiosRequestConfig; }): RestResponse<any> {
+    request<R>(requestConfig: { method: string; url: string; queryParams?: any; data?: any; copyFn?: (data: R) => R; options?: Axios.AxiosRequestConfig; }): RestResponse<R> {
         function assign(target: any, source?: any) {
             if (source != undefined) {
                 for (const key in source) {
@@ -31,8 +31,15 @@ class AxiosHttpClient implements HttpClient<Axios.AxiosRequestConfig> {
         config.params = requestConfig.queryParams;
         config.data = requestConfig.data;
         assign(config, requestConfig.options);
+        const copyFn = requestConfig.copyFn;
 
         const axiosResponse = this.axios.request(config);
-        return axiosResponse;
+        return axiosResponse.then(axiosResponse => {
+            if (copyFn && axiosResponse.data) {
+                (axiosResponse as any).originalData = axiosResponse.data;
+                axiosResponse.data = copyFn(axiosResponse.data);
+            }
+            return axiosResponse;
+        });
     }
 }
