@@ -1,7 +1,7 @@
 package cz.habarta.typescript.generator;
 
-import static org.junit.Assert.*;
 import org.junit.*;
+import static org.junit.Assert.*;
 
 public class SortedTypesTest {
 
